refactor netty pipeline
use single message codec with state
